requirements.txt
pynput==1.7.6
python-dotenv==1.0.0
pywin32==306
